Remove C99 style for loop
Several for loops have `for (int n=0; ...)` which requires C99 or later compilation.  However, the `n` variable is already defined in these routines, so the declaration isn't even needed.  Just changing `int n` to `n` in the for loops removes need for C99 and also doesn't shadow the outer `n`.